+++ CHANGES.md
@@ -12,6 +12,8 @@ Unreleased
 - Allow %{read-lines} to introduce dynamic dependencies like %{read}. (#5440,
   @anmonteiro)
 
+- Look up `gmake` before `make` (#.., fixes #5470, @rgrinberg)
+
 3.0.2 (17/02/2022)
 ------------------